Remove lteOnCdmaDevice and set preferred network on a per-device basis
Not all of our devices are CDMA-capable.

While the old configuration "worked" it was nonoptimal:
1) When SIM is removed and in airplane mode, the device should not have anything on lockscreen.
2) Due to running CdmaLteSST, the device reports a PLMN as "No SIM Card" in this configuration.
3) Likely to cause strange preferred network mode management issues.

Thanks to libinit, it is better to properly configure each device for the network type
it is intended for.

This fixes the issue of "No SIM Card" or "No Service" appearing on the lockscreen of
GSM devices when a SIM is removed and the device is in airplane mode.  (These messages
should only appear when the GSM radio is active.)